requirements.txt
python-telegram-bot==20.6
ultralytics
opencv-python
numpy
speechrecognition
pydub